[refacotr] 댓글 UI 리팩토링
