+++ dutchiepay/src/app/_components/_community/_free/_comment/CommentForm.jsx
@@ -1,16 +1,12 @@
 'use client';
 
-import '@/styles/globals.css';
-
-import CommentWrite from './CommentWrite';
+import CommentSubmit from './CommentSubmit';
 import FreeCommentList from './FreeCommentList';
 import Image from 'next/image';
 import communityComment from '/public/image/community/communityComment.svg';
-import profile from '/public/image/profile.jpg';
 import useInfiniteScroll from '@/app/hooks/useInfiniteScroll';
-import { useSelector } from 'react-redux';
 
-const CommentForm = ({ postId, post }) => {
+export default function CommentForm({ postId, post }) {
   const fetchUrl = `${process.env.NEXT_PUBLIC_BASE_URL}/free/comments/list?freeId=${postId}&limit=6`;
   const {
     items: comments,
@@ -18,7 +14,6 @@ const CommentForm = ({ postId, post }) => {
     lastItemRef,
     refresh: refreshComments,
   } = useInfiniteScroll({ fetchUrl });
-  const profileImage = useSelector((state) => state.login.user.profileImage);
 
   return (
     <div className="mt-[40px]">
@@ -26,21 +21,11 @@ const CommentForm = ({ postId, post }) => {
         <h2 className="text-xl font-bold">댓글</h2>
         <p>{post.commentsCount}개</p>
       </div>
-      <div className="flex justify-between my-[12px]">
-        <div className="relative w-[50px] h-[50px] rounded-full border">
-          <Image
-            className="w-[50px] h-[50px] rounded-full object-cover"
-            src={profileImage || profile}
-            alt="프로필"
-            fill
-          />
-        </div>
-        <CommentWrite
-          postId={postId}
-          refreshComments={refreshComments}
-          type={'comment'}
-        />
-      </div>
+      <CommentSubmit
+        postId={postId}
+        refreshComments={refreshComments}
+        type={'comment'}
+      />
       {!isInitialized || comments.length === 0 ? (
         <div className="mx-auto my-auto  flex flex-col justify-center items-center">
           <Image
@@ -50,14 +35,14 @@ const CommentForm = ({ postId, post }) => {
             height={58}
             className="mt-[60px] mb-[12px] mx-auto"
           />
-          <p className="text-sm text-center mb-[50px]">
+          <p className="text-sm text-center mb-[50px] text-gray--500">
             현재 등록된 댓글이 없습니다.
             <br />
             새로운 댓글을 작성하여 다양한 의견과 정보를 공유해 주세요.
           </p>
         </div>
       ) : (
-        <div className="border-b py-[16px]">
+        <div className="mt-[24px] border-t py-[16px]">
           {comments.map((item, index) => (
             <FreeCommentList
               key={index}
@@ -72,6 +57,4 @@ const CommentForm = ({ postId, post }) => {
       )}
     </div>
   );
-};
-
-export default CommentForm;
+}
